fixed velocity arithmetic

+++ Programs/Drehgeber/angle.h
@@ -5,7 +5,7 @@
 #define ANGLE_PER_PHASE 0.3
 
 
-double getVelocity(int phaseDiff, int deltaTime);
+double getVelocity(int phaseDiff, double deltaTime);
 double calculateAngle(int steps);
 
 #endif /* ANGLE_H */
